+++ R/develop_rdf_snomed.R
@@ -1,19 +1,20 @@
-develop_rdf_rxnorm_atc <-
-  function(conn,
-           conn_fun = "pg13::local_connect()",
-           schema = "omop_athena",
-           output_folder = "~/Desktop",
-           checks = "",
-           verbose = TRUE,
-           render_sql = TRUE,
-           render_only = FALSE) {
+develop_rdf_snomed <-
+  function(
+    conn,
+    conn_fun = "pg13::local_connect()",
+    schema = "omop_athena",
+    output_folder = "~/Desktop",
+    checks = "",
+    verbose = TRUE,
+    render_sql = TRUE,
+    render_only = FALSE) {
 
 
 omop_version <-
   pg13::query(
     conn_fun = conn_fun,
     sql_statement =
-    "
+      "
     SELECT sa_release_version
     FROM public.setup_athena_log
     WHERE sa_datetime IN (SELECT MAX(sa_datetime) FROM public.setup_athena_log);
@@ -26,23 +27,30 @@ omop_version <-
 
 
 rdf_file <-
-file.path(
-"inst",
-"data",
-"rdf",
-omop_version,
-"Drug",
-"RxNorm ATC Classification.rdf"
- )
+  file.path(
+    "inst",
+    "data",
+    "rdf",
+    omop_version,
+    "Condition",
+    "SNOMED.rdf"
+  )
 
 if (!file.exists(rdf_file)) {
-# Workflow
+
+
 library(tidyverse)
+library(easyBakeOven)
+library(glue)
+
+# Summary
+# Convert the ATC and RxNorm classification in the OMOP vocabularies to RDF.
+# Workflow
 library(rdflib)
 library(rdfR)
 library(pg13)
-conn_fun <- "pg13::local_connect()"
 
+conn_fun <- "pg13::local_connect()"
 get_athena_version <-
   function(conn,
            conn_fun,
@@ -69,11 +77,13 @@ omop_version <-
   get_athena_version(conn_fun = conn_fun)
 omop_version
 
+
+
 drug_rdf <-
-  initialize_rdf(name = "OMOP RxNorm ATC Classification",
-                 base_uri = "http://omop/drug",
-                 version =  omop_version,
-                 comment = "ATC classes with RxNorm and RxNorm Extension Ingredient concept class individuals.")
+  initialize_rdf(name = "OMOP SNOMED Ontology",
+                 base_uri = "http://omop/snomed/condition",
+                 version = omop_version$version,
+                 comment = "SNOMED Condition domain ontology as it is available at athena.ohdsi.org.")
 
 
 base_uri <- query_base_uri(drug_rdf)
@@ -85,33 +95,39 @@ drug_rdf <-
 
 
 add_annotation_property_value(rdf = drug_rdf,
-                            entity = base_uri,
-                            annotation_property = "load_datetime",
-                            value = omop_version$load_datetime)
+                              entity = base_uri,
+                              annotation_property = "load_datetime",
+                              value = omop_version$load_datetime)
 
 add_annotation_property_value(rdf = drug_rdf,
-                            entity = base_uri,
-                            annotation_property = "version",
-                            value = omop_version$version)
+                              entity = base_uri,
+                              annotation_property = "version",
+                              value = omop_version$version)
 drug_rdf
 
 
-# Reading OMOP ATC classifications.
-atc_classes <-
-query(sql_statement = glue::glue("SELECT * FROM {schema}.concept WHERE vocabulary_id = 'ATC' AND standard_concept = 'C' AND invalid_reason IS NULL;"))
-atc_classes
 
-atc_classes2 <-
-  atc_classes %>%
+
+classes <-
+query(sql_statement = glue::glue("SELECT * FROM {schema}.concept WHERE vocabulary_id = 'ICDO3' AND domain_id = 'Condition' AND invalid_reason IS NULL;"))
+classes
+
+
+
+classes2 <-
+  classes %>%
   mutate(class_id = sprintf("%s/%s", base_uri,concept_id)) %>%
   mutate(class_label = sprintf("%s [%s %s]", concept_name, vocabulary_id, concept_code))
-atc_classes2
+
+classes2
+
+
 
 drug_rdf2 <-
   append_annotation_properties(
     rdf = drug_rdf,
     base_uri = base_uri,
-    colnames(atc_classes)
+    colnames(classes)
   )
 drug_rdf2
 
@@ -120,7 +136,7 @@ drug_rdf2
 drug_rdf3 <-
 df_add_classes(
   rdf = drug_rdf2,
-  data = atc_classes2,
+  data = classes2,
   class_id_col = class_id,
   class_label_col = class_label
 )
@@ -128,92 +144,105 @@ drug_rdf3
 
 
 
-atc_classes3 <-
-atc_classes2 %>%
+classes3 <-
+classes2 %>%
   select(-class_label) %>%
   mutate_all(as.character) %>%
   pivot_longer(cols = !class_id,
                names_to = "annotation_property_label",
                values_drop_na = TRUE)
 
-atc_classes3
+classes3
 
 
 
-atc_classes4 <-
-  atc_classes3 %>%
+classes4 <-
+  classes3 %>%
   split(.$annotation_property_label) %>%
   map(select,
       -annotation_property_label)
 
-atc_classes4
+classes4
 
 
 
 drug_rdf4 <-
   drug_rdf3
-for (i in seq_along(atc_classes4)) {
+for (i in seq_along(classes4)) {
   drug_rdf4 <-
   df_add_annotation_property_value(
     rdf = drug_rdf4,
-    data = atc_classes4[[i]],
+    data = classes4[[i]],
     entity_col = class_id,
-    annotation_property_label = names(atc_classes4)[i],
+    annotation_property_label = names(classes4)[i],
     value_col = value
   )
 }
 
 
 
+subclassof <-
+query(sql_statement = glue::glue("SELECT ca.* FROM {schema}.concept_ancestor ca INNER JOIN {schema}.concept a ON a.concept_id = ca.ancestor_concept_id INNER JOIN {schema}.concept d ON d.concept_id = ca.descendant_concept_id WHERE a.vocabulary_id = 'HemOnc' AND a.standard_concept = 'C' AND a.invalid_reason IS NULL AND d.vocabulary_id = 'HemOnc' AND d.standard_concept = 'C' AND d.invalid_reason IS NULL;"))
+subclassof
 
-atc_subclassof <-
-query(sql_statement = glue::glue("SELECT ca.* FROM {schema}.concept_ancestor ca INNER JOIN {schema}.concept a ON a.concept_id = ca.ancestor_concept_id INNER JOIN {schema}.concept d ON d.concept_id = ca.descendant_concept_id WHERE a.vocabulary_id = 'ATC' AND a.standard_concept = 'C' AND a.invalid_reason IS NULL AND D.vocabulary_id = 'ATC' AND d.standard_concept = 'C' AND d.invalid_reason IS NULL;"))
-atc_subclassof
 
 
-atc_subclassof %>%
+subclassof %>%
   dplyr::filter(ancestor_concept_id == descendant_concept_id) %>%
   distinct(min_levels_of_separation, max_levels_of_separation)
 
 
-atc_subclassof2 <-
-  atc_subclassof %>%
+
+subclassof2 <-
+  subclassof %>%
   dplyr::filter(ancestor_concept_id != descendant_concept_id)
-atc_subclassof2
+subclassof2
 
 
 
-atc_subclassof2 %>%
-  dplyr::filter(min_levels_of_separation != max_levels_of_separation)
+subclassof2 %>%
+  dplyr::filter(min_levels_of_separation != max_levels_of_separation) %>%
+  distinct(min_levels_of_separation,
+           max_levels_of_separation)
 
 
 
-atc_subclassof3 <-
-  atc_subclassof2 %>%
+subclassof2 %>%
+  count(min_levels_of_separation)
+
+subclassof2 %>%
+  count(max_levels_of_separation)
+
+
+
+
+subclassof3 <-
+  subclassof2 %>%
   transmute(ancestor_concept_id,
             descendant_concept_id,
             level = min_levels_of_separation)
-atc_subclassof3
+subclassof3
 
 
 
-atc_subclassof4 <-
-  atc_subclassof3 %>%
+subclassof4 <-
+  subclassof3 %>%
   mutate_at(vars(ancestor_concept_id,
                  descendant_concept_id),
             ~sprintf("%s/%s", base_uri,.))
-atc_subclassof4
+subclassof4
 
 
 
-atc_subclassof5 <-
-  atc_subclassof4 %>%
+subclassof5 <-
+  subclassof4 %>%
   pivot_wider(names_from = level,
               values_from = descendant_concept_id,
               values_fn = list) %>%
   select(ancestor_concept_id,
          as.character(1:4))
-atc_subclassof5
+subclassof5
+
 
 
 drug_rdf5 <- drug_rdf4
@@ -234,18 +263,18 @@ for (i in 1:4) {
   }
 
 
-  atc_subclassof6 <-
-  atc_subclassof5 %>%
+  subclassof6 <-
+  subclassof5 %>%
     select(all_of(c(col1, col2))) %>%
     unnest(all_of(col2))
 
-  colnames(atc_subclassof6) <-
+  colnames(subclassof6) <-
     c("parent_id", "child_id")
 
   drug_rdf5 <-
   df_add_subclassof(
     rdf = drug_rdf5,
-    data = atc_subclassof6,
+    data = subclassof6,
     parent_class_id_col = parent_id,
     child_class_id_col = child_id
   )
@@ -256,18 +285,18 @@ drug_rdf5
 
 
 
-rxnorm_individuals <-
-query(sql_statement = glue::glue("SELECT * FROM {schema}.concept WHERE vocabulary_id IN ('RxNorm', 'RxNorm Extension') AND standard_concept <> 'C' AND concept_class_id = 'Ingredient' AND invalid_reason IS NULL;"))
-rxnorm_individuals
+individuals <-
+query(sql_statement = glue::glue("SELECT * FROM {schema}.concept WHERE vocabulary_id IN ('HemOnc') AND standard_concept <> 'C' AND  invalid_reason IS NULL;"))
+individuals
 
 
 
-rxnorm_individuals2 <-
-  rxnorm_individuals %>%
+individuals2 <-
+  individuals %>%
   mutate(individual_id = sprintf("%s/%s", base_uri,concept_id)) %>%
   mutate(individual_label = sprintf("%s [%s %s]", concept_name, vocabulary_id, concept_code))
 
-rxnorm_individuals2
+individuals2
 
 
 
@@ -276,7 +305,7 @@ drug_rdf6 <- drug_rdf5
 drug_rdf6 <-
 df_add_individuals(
   rdf = drug_rdf6,
-  data = rxnorm_individuals2,
+  data = individuals2,
   individual_id_col = individual_id,
   individual_label_col = individual_label
 )
@@ -284,44 +313,43 @@ df_add_individuals(
 
 
 
-rxnorm_individuals3 <-
-rxnorm_individuals2 %>%
+individuals3 <-
+individuals2 %>%
   select(-individual_label) %>%
   mutate_all(as.character) %>%
   pivot_longer(cols = !individual_id,
                names_to = "annotation_property_label",
                values_drop_na = TRUE)
 
-rxnorm_individuals3
+individuals3
 
 
 
-rxnorm_individuals4 <-
-  rxnorm_individuals3 %>%
+individuals4 <-
+  individuals3 %>%
   split(.$annotation_property_label) %>%
   map(select,
       -annotation_property_label)
 
-rxnorm_individuals4
+individuals4
 
 
 
 drug_rdf7 <- drug_rdf6
-for (i in seq_along(rxnorm_individuals4)) {
+for (i in seq_along(individuals4)) {
   drug_rdf7 <-
   df_add_annotation_property_value(
     rdf = drug_rdf7,
-    data = rxnorm_individuals4[[i]],
+    data = individuals4[[i]],
     entity_col = individual_id,
-    annotation_property_label = names(rxnorm_individuals4)[i],
+    annotation_property_label = names(individuals4)[i],
     value_col = value
   )
 }
 
 
 
-
-rxnorm_individual_classes <-
+individual_classes <-
 query(
   sql_statement =
     glue::glue(
@@ -332,9 +360,8 @@ query(
     (SELECT concept_id
     FROM {schema}.concept rx
     WHERE
-      rx.vocabulary_id IN ('RxNorm', 'RxNorm Extension') AND
+      rx.vocabulary_id IN ('HemOnc') AND
       rx.standard_concept <> 'C' AND
-      rx.concept_class_id = 'Ingredient' AND
       rx.invalid_reason IS NULL) c
   ON c.concept_id = ca.descendant_concept_id
   INNER JOIN
@@ -342,30 +369,30 @@ query(
     SELECT concept_id
     FROM {schema}.concept
     WHERE
-      vocabulary_id IN ('ATC') AND
+      vocabulary_id IN ('HemOnc') AND
       standard_concept = 'C' AND
       invalid_reason IS NULL
     ) atc
   ON atc.concept_id = ca.ancestor_concept_id;"))
-rxnorm_individual_classes
+individual_classes
 
 
 
-rxnorm_individual_classes %>%
+individual_classes %>%
   distinct(min_levels_of_separation,
            max_levels_of_separation)
 
 
 
-rxnorm_individual_classes %>%
+individual_classes %>%
   dplyr::filter(min_levels_of_separation == 0,
                 max_levels_of_separation == 0)
 
 
 
 
-rxnorm_individual_classes2 <-
-  rxnorm_individual_classes %>%
+individual_classes2 <-
+  individual_classes %>%
   transmute(ancestor_concept_id,
             descendant_concept_id,
             level = max_levels_of_separation) %>%
@@ -373,7 +400,7 @@ rxnorm_individual_classes2 <-
   select(ancestor_concept_id,
          descendant_concept_id) %>%
   mutate_all(~sprintf("%s/%s", base_uri, .))
-rxnorm_individual_classes2
+individual_classes2
 
 
 
@@ -381,20 +408,23 @@ drug_rdf8 <- drug_rdf7
 drug_rdf8 <-
 df_add_individual_class(
   rdf = drug_rdf8,
-  data = rxnorm_individual_classes2,
+  data = individual_classes2,
   individual_id_col = descendant_concept_id,
   class_id_col = ancestor_concept_id
 )
 
 
 
-
 final_rdf <-
   drug_rdf8
 
+write_rdf(
+  final_rdf,
+  rdf_file
+)
 
-write_rdf(final_rdf,
-  rdf_file)
 }
 
 }
+
+